whisper-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
torch==2.1.0
numpy<2.0.0

